+++ spec/models/identifier/local/record_number_spec.rb
@@ -3,13 +3,18 @@ describe Identifier::Local::RecordNumber, type: :model, group: :identifiers do
 
   let(:n) { FactoryBot.create(:valid_namespace, short_name: 'NS', delimiter: nil) }
 
-  specify 'assigned to CollectionObject only' do
+  specify 'assigned to invalid object' do
     i = Identifier::Local::RecordNumber.create(identifier: '345', namespace: n, identifier_object: FactoryBot.create(:valid_collecting_event))
     expect(i.errors.key?(:identifier_object_type)).to be_truthy
   end
 
-  specify 'assigned to CollectionObject only 2' do
+  specify 'assigned to CollectionObject' do
     i = Identifier::Local::RecordNumber.create(identifier: '345', namespace: n, identifier_object: Specimen.create!)
+    expect(i.errors.key?(:identifier_object_type)).to be_falsey
+  end
+
+  specify 'assigned to FieldOccurrence' do
+    i = Identifier::Local::RecordNumber.create(identifier: '345', namespace: n, identifier_object: FactoryBot.create(:valid_field_occurrence))
     expect(i.errors.key?(:identifier_object_type)).to be_falsey
   end
 
@@ -19,4 +24,9 @@ describe Identifier::Local::RecordNumber, type: :model, group: :identifiers do
     expect(j.valid?).to be_truthy
   end
 
+  specify 'may be duplicated across field occurrences' do
+    i = Identifier::Local::RecordNumber.create!(identifier: '345', namespace: n, identifier_object: FactoryBot.create(:valid_field_occurrence))
+    j = Identifier::Local::RecordNumber.new(identifier: '345', namespace: n, identifier_object: FactoryBot.create(:valid_field_occurrence))
+    expect(j.valid?).to be_truthy
+  end
 end