+++ CHANGELOG.md
@@ -7,8 +7,25 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 
 ## [unreleased]
 
+### Added
+
+- Preview material examined task [#2325]
+
+### Changed
+
+- Allow catalog and record number identifiers on field occurrences
+
 ### Fixed
-Taxon Pages Biological Associations returns no results
+
+- Taxon Pages Biological Associations returns no results
+- error on /api/v1/dwc_occurrences
+- /downloads/1 shows `Expires: n days` when the download has been *expired* for n days
+- New Source can fail to update a serial when the serial being updated to has no issn [#4921]
+- In New Source, an issn value never overrides the serial selected by the user
+
+[#4921]: https://github.com/SpeciesFileGroup/taxonworks/issues/4921
+
+[#2325]: https://github.com/SpeciesFileGroup/taxonworks/issues/2325
 
 ## [0.62.0] - 2026-05-08