+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 
 - Preview material examined task [#2325]
 - Add Filter CO/FO to CE radial navigator [#4925]
+- `TW:TaxonDetermination:otu_id` DwC occurrence importer field for direct OTU mapping
 
 ### Changed
 
@@ -25,7 +26,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - New Source can fail to update a serial when the serial being updated to has no issn [#4921]
 - In New Source, an issn value never overrides the serial selected by the user
 - Filter CO: Biological associations facet doesn't work with the `without` option [#4926]
- Sqed images are automatically processed when created [#3178]
+- Sqed images are automatically processed when created [#3178]
 
 [#4921]: https://github.com/SpeciesFileGroup/taxonworks/issues/4921
 [#2325]: https://github.com/SpeciesFileGroup/taxonworks/issues/2325