+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 ### Added
 
 - Preview material examined task [#2325]
+- Unify projects (a rake task, not runnable from the UI) [#4445]
 - Add Filter CO/FO to CE radial navigator [#4925]
 - `TW:TaxonDetermination:otu_id` DwC occurrence importer field for direct OTU mapping
 
@@ -30,6 +31,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 
 [#4921]: https://github.com/SpeciesFileGroup/taxonworks/issues/4921
 [#2325]: https://github.com/SpeciesFileGroup/taxonworks/issues/2325
+[#4445]: https://github.com/SpeciesFileGroup/taxonworks/issues/4445
 [#3178]: https://github.com/SpeciesFileGroup/taxonworks/issues/3178
 [#4925]: https://github.com/SpeciesFileGroup/taxonworks/issues/4925
 [#4926]: https://github.com/SpeciesFileGroup/taxonworks/issues/4926